pyproject.toml
[project]
name = "efscape"
version = "1.0.0"
description = "Implements a modeling and simulation (M&S) framework that utilizes ZeroC Ice RPC middleware to deploy a M&S client-server application"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Jon C Cline", email = "clinej@alumni.stanford.edu" }]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "xdevs>=3.0.0",
    "click>=8.1.8",
    "click-log>=0.4.0",
    "python-dotenv>=1.0.1",
    "pandas>=2.2.3",
    "zeroc-ice>=3.7.10.1",
    "pyarrow>=18.1.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.10.0",
    "ruff>=0.8.4",
    "pytest>=8.3.4",
    "nox>=2024.10.9",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["efscape*", "arrowice*"]
